admin_backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic[email]==2.11.7
python-dotenv==1.0.1
PyJWT==2.9.0
